refactor(web_server): create sane defaults for testing
Create a create_app_for_testing function which uses a dummy AuthDb by
default instead of the production AuthDb. This doesn't change anything,
as tests were not using create_app's default AuthDb configuration, but I
think it's nice to pick better defaults for tests to prevent issues in
the future.